feat(size): add code improvements
