+++ src/models/layer-model/common/data-model/utils/__tests__/data-utils.spec.ts
@@ -1,4 +1,5 @@
 import DataUtils from '../data-utils';
+import mockLayout from '../../../../../../mocks/layout';
 
 describe('DataUtils', () => {
   describe('getAttributeData', () => {
@@ -72,6 +73,22 @@ describe('DataUtils', () => {
       const value = { qNum: 'NaN' };
       const attrExpr = DataUtils.extractAttributeExpression(value);
       expect(attrExpr).toBe(null);
+    });
+  });
+
+  describe('flattenDataPages', () => {
+    let dataPages: NxDataPage[];
+
+    beforeEach(() => {
+      dataPages = mockLayout.layer.base.qHyperCube.qDataPages;
+    });
+    it('should flatten data', () => {
+      dataPages.push({
+        qMatrix: [[{ qElemNumber: 0, qState: 'S' }], [{ qElemNumber: 1, qState: 'S' }]],
+      });
+      const result = DataUtils.flattenDataPages(dataPages);
+      expect(result.qMatrix).toBeDefined();
+      expect(result.qMatrix?.length).toBe(3);
     });
   });
 });
