+++ src/models/layer-model/common/data-model/utils/data-utils.ts
@@ -32,6 +32,22 @@ module DataUtils {
     }
     return null;
   }
+
+  /**
+   * Takes a NxDataPage[] and transforms it into one big NxDataPage.
+   * @param data Array to be flattened.
+   */
+  export function flattenDataPages(data: NxDataPage[]): NxDataPage {
+    const newDataPage: NxDataPage = data[0];
+    const flattened: NxCell[][] = [];
+    newDataPage.qMatrix = flattened.concat.apply(
+      [],
+      data.map((row) => {
+        return row.qMatrix as NxCell[][];
+      })
+    );
+    return newDataPage;
+  }
 }
 
 export default DataUtils;
