+++ src/typings/data/data.d.ts
@@ -23,9 +23,3 @@ declare interface LocationData {
 declare interface PointData extends Data {
   qAttrExps?: NxAttributeExpressionValues;
 }
-
-declare interface DimensionExpressionInfo {
-  index: number;
-  isDim: bolean;
-  dimensionIndex: number;
-}
